Move out duration input
